+++ src/bin/start.ts
@@ -1,7 +1,6 @@
 import "reflect-metadata";
 import { container } from "tsyringe";
-import { configureMikro, knexConfig, KnexRef } from "../db";
-import Knex from "knex";
+import { configureMikro } from "../db";
 import { CreateAPI } from "../api";
 import { serve } from "@hono/node-server";
 import { Configuration } from "../config";
@@ -9,17 +8,14 @@ import { BaseLogger } from "../util/logger";
 import { MikroORM } from "@mikro-orm/postgresql";
 
 (async () => {
-    const knex = Knex(knexConfig);
-    container.register(KnexRef, { useValue: knex });
-
-    const mikroORM = await configureMikro(knex);
+    const mikroORM = await configureMikro();
     container.register(MikroORM, { useValue: mikroORM });
 
     const log = container.resolve(BaseLogger);
 
     const app = CreateAPI(container);
 
-    await knex.migrate.latest();
+    await mikroORM.migrator.up();
 
     serve(
         {
